backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Kal Cantrell <klcantrell@sep.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.112.0"}
langchain-ollama = "^0.1.1"
langchain-core = "^0.2.28"
typing-extensions = "^4.12.2"
langchain-openai = "^0.1.22"
python-dotenv = "^1.0.1"
langgraph = "^0.2.14"
tavily-python = "^0.4.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.11.1"
ruff = "^0.5.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.format]
quote-style = "double"
